refactor: add set-settings file to set-up default language for keyboard
