+++ js/app.js
@@ -1,24 +1,22 @@
 import createTextField from './components/create-text-field.js';
 import createElement from './components/create-element.js';
 import setTextValue from './components/set-text-value.js';
-import createKeyboard from './components/create-keyboard.js';
 import langs from './data/langs.js';
 import switchLang from './components/switch-lang.js';
+import setSettings from './components/set-settings.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   const container = createElement('div', 'container');
   const keyboard = createElement('div', 'keyboard');
-  let keyboardContainer = createElement('div', 'keyboard__container');
   const field = createTextField();
 
-  document.body.append(container);
+  setSettings(langs(), keyboard);
 
-  keyboardContainer.append(...createKeyboard(langs()));
-  keyboard.append(keyboardContainer);
+  document.body.append(container);
   container.append(field);
   container.append(keyboard);
+  switchLang(langs(), keyboard);
 
   keyboard.addEventListener('click', setTextValue);
-  switchLang(langs(), keyboardContainer);
 });
