+++ js/components/create-buttons.js
@@ -2,7 +2,7 @@ import createElement from './create-element.js';
 import checkSymbol from './check-symbol.js';
 import checkKey from './check-key.js';
 
-export default function createKeyboardRow(lang, keys) {
+export default function createButtons(lang, keys) {
   const buttons = [];
 
   for (let i = 0; i < keys.length; i++) {
