+++ js/components/create-keyboard.js
@@ -1,10 +1,10 @@
-import createArrayRows from './create-array-rows.js';
+import createRows from './create-rows.js';
 import createElement from './create-element.js';
 
 export default function createKeyboard(lang) {
   let rowsContainer = [];
 
-  createArrayRows(lang).forEach((rows) => {
+  createRows(lang).forEach((rows) => {
     let row = createElement('div', 'keyboard__row');
     row.append(...rows);
     rowsContainer.push(row);
