+++ js/components/set-settings.js
@@ -1,0 +1,9 @@
+import createKeyboard from './create-keyboard.js';
+
+export default function setSettings(lang, keyboard) {
+  keyboard.innerHTML = '';
+
+  createKeyboard(lang).forEach((row) => {
+    keyboard.innerHTML += row.outerHTML;
+  });
+};
