+++ js/components/switch-lang.js
@@ -1,16 +1,12 @@
+import setSettings from './set-settings.js';
 import changeLang from './change-lang.js';
-import createKeyboard from './create-keyboard.js';
 
 export default function switchLang(lang, keyboard) {
 
   document.addEventListener('keydown', (e) => {
     if (e.altKey && e.ctrlKey) {
       lang = changeLang(lang);
-      keyboard.innerHTML = '';
-
-      createKeyboard(lang).forEach((row) => {
-        keyboard.innerHTML += row.outerHTML;
-      });
+      setSettings(lang, keyboard);
     }
   });
